+++ packages/builder-vue3/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @varlinor/builder-vue3
 
+## 1.0.7
+
+### Patch Changes
+
+- feat: modify generate entry in cli
+
+## 1.0.6
+
+### Patch Changes
+
+- refactor: split vue2 builder and vue3 builder
+
+## 1.0.6-snapshot.0
+
+### Patch Changes
+
+- refactor: split vue2 builder and vue3 builder
+
 ## 1.0.5
 
 ### Patch Changes