+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @varlinor/cli
 
+## 1.0.5
+
+### Patch Changes
+
+- feat: modify generate entry in cli
+
 ## 1.0.4
 
 ### Patch Changes